Fix race between OffThreadRandomizer::ExecuteOnFile and app exit
There was a race condition between OffThreadRandomizer::ExecuteOnFile
and appliation exit: it was possible for the OffThreadRandomizer to
try to connect to the MainWindow after the MainWindow had already
been destroyed and its memory freed.  This prevents that from
happening by having the MainWindow destructor wait for the future
returned by QtConcurrent::run to complete, which also ensures that the
randomizer will be allowed to complete once started, unless the
program is killed by a signal or the like.